requirements.txt
requests
openai
beautifulsoup4
lxml
python-magic
thefuzz
trafilatura
